requirements.txt
python-telegram-bot>=20.0
fastapi>=0.100
uvicorn>=0.23
gspread>=5.10
google-auth>=2.20
python-dotenv>=1.0
